refactor: major cleanup and restructuring
This refactor regards the following changes:
- The main code has been moved to the `App`
struct, which will now handle everything.
This is useful because it makes it more readable
where the entry point is.
- Algebra (Vec3, Vec2, Vertices) live in a different
file now (algebra.rs), as opposed to the larger
`shape.rs` file. More is to be done regarding this.

Future TODO: Refactor everything into
a separate crate regarding the maths.