feat: primitive BVH Tree
This feature speeds up rendering by a *lot*.
It does this by improving the way we check for intersections.
Previously, we would go through each and every candidate
(which means everything!) and check if they intersect. Now,
we can filter out a lot of candidates just by the fact that
we use bounding boxes.

TODO: We must improve the performance by sorting the candidates
in a better way.